+++ releases.md
@@ -9,7 +9,7 @@ bleeding edge...
 * jexler-core
   * Slightly modified interface of ShellTool (lists and maps
     instead of arrays to make it simpler to use with Groovy).
-  * More unit tests (jacoco XX%/YY%).
+  * More unit test coverage.
 * ...
 
 1.0.3 (2013-07-16)